add 4th option to query keys internally
due to amount of parameters used for internal querying i
decided to switch to a params-struct. this is the starting point of a
larger refactoring to simplify the `I*ProjectionStore` interfaces.
this will probably make the implementation of these interfaces more
complex, but that can be simplified in the future